Buffer management refactoring (#28)

+++ DnsCore/Model/Encoding/DnsMessageEncoder.cs
@@ -4,5 +4,5 @@ namespace DnsCore.Model.Encoding;
 
 public static class DnsMessageEncoder
 {
-    public static int Encode(Span<byte> buffer, DnsMessage message) => DnsRawMessageEncoder.Encode(buffer, message.ToRawMessage());
+    public static ushort Encode(Span<byte> buffer, DnsMessage message) => DnsRawMessageEncoder.Encode(buffer, message.ToRawMessage());
 }
